# spec/openapi.yaml
---
openapi: 3.0.3
info:
  title: Kamifusen API
  description: Page hit counter administration API
  version: 0.9.0
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
paths:
  /admin/keygen:
    post:
      tags:
      - App Admin Resource
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
              - username
              - role
              - expiresAt
              type: object
              properties:
                username:
                  type: string
                role:
                  type: string
                expiresAt:
                  type: string
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /admin/pagedel/{pageId}:
    post:
      tags:
      - App Admin Resource
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            text/plain: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /admin/pages:
    get:
      tags:
      - App Admin Resource
      responses:
        "200":
          description: List of registered pages with visit statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PageWithStatsDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /admin/retire/{userId}:
    post:
      tags:
      - App Admin Resource
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /admin/stats:
    get:
      tags:
      - App Admin Resource
      parameters:
      - name: timeRange
        in: query
        schema:
          type: string
          nullable: true
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregatedStatsDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /admin/users:
    get:
      tags:
      - App Admin Resource
      responses:
        "200":
          description: List of API users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiUser"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /admin/visits:
    get:
      tags:
      - App Admin Resource
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /public/visits/count/{pageId}:
    get:
      tags:
      - Page Visit Resource
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /public/visits/hit:
    post:
      tags:
      - Page Visit Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageHitRequestDto"
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
components:
  schemas:
    AggregatedStatsDto:
      required:
      - visitData
      - topPages
      - domainStats
      type: object
      properties:
        visitData:
          type: array
          items:
            $ref: "#/components/schemas/VisitTrendDataDto"
        topPages:
          type: array
          items:
            $ref: "#/components/schemas/TopPageDataDto"
        domainStats:
          type: array
          items:
            $ref: "#/components/schemas/DomainStatDataDto"
        totalVisits:
          format: int64
          type: integer
        totalPages:
          format: int64
          type: integer
        totalDomains:
          format: int64
          type: integer
    ApiUser:
      required:
      - username
      - role
      type: object
      properties:
        id:
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
          nullable: true
        username:
          type: string
        password:
          type: string
          nullable: true
        role:
          type: string
        expiresAt:
          type: string
          allOf:
          - $ref: "#/components/schemas/LocalDateTime"
          nullable: true
        added:
          type: string
          allOf:
          - $ref: "#/components/schemas/LocalDateTime"
          nullable: true
        updated:
          type: string
          allOf:
          - $ref: "#/components/schemas/LocalDateTime"
          nullable: true
    DomainStatDataDto:
      required:
      - domain
      - percentage
      type: object
      properties:
        domain:
          type: string
        visits:
          format: int64
          type: integer
        percentage:
          type: number
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
    PageHitRequestDto:
      required:
      - path
      - domain
      type: object
      properties:
        path:
          type: string
        domain:
          type: string
    PageWithStatsDto:
      required:
      - id
      - path
      - pageAdded
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        path:
          type: string
        domain:
          type: string
          nullable: true
        pageAdded:
          $ref: "#/components/schemas/LocalDateTime"
        lastHit:
          type: string
          allOf:
          - $ref: "#/components/schemas/LocalDateTime"
          nullable: true
        visitCount:
          format: int64
          type: integer
    TopPageDataDto:
      required:
      - path
      - percentage
      type: object
      properties:
        path:
          type: string
        visits:
          format: int64
          type: integer
        percentage:
          type: number
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
    VisitTrendDataDto:
      required:
      - label
      - category
      type: object
      properties:
        label:
          type: string
        value:
          format: int64
          type: integer
        category:
          type: string
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: basic